Add files via upload
I created a GUI for the Grocery store list. This is my first time using the tkinter module. In this first program, the user inputs their total for each item an the program and when the user clicks the check out button, the total is displayed in the GUI. As this was my first time using tkinter, it is not the most visually pleasing application but now that I could the basic of tkinter down I am going to improve the visual as well as look into adding other features (ie checkbox for delivery option that will add delivery fee to total)